+++ src/pages/project/ProjectRequest/ProjectRequestCompleteSection.tsx
@@ -5,7 +5,7 @@ import { Button } from "principes-getp";
 import { Paragraph } from "@/components/__common__/typography/Paragraph";
 import { Title } from "@/components/__common__/typography/Title";
 
-import { ProjectRequestPageWrapper } from "@/pages/project/ProjectRequestPage.style";
+import { ProjectRequestPageWrapper } from "@/pages/project/ProjectRequest/ProjectRequestPage.style";
 
 import congratsImage from "@/assets/auth/congrats1.png";
